gemini generated new_eligibility.html
