+++ config/urls.py
@@ -8,7 +8,8 @@ from django.views.generic import TemplateView
 from drf_spectacular.views import SpectacularAPIView
 from drf_spectacular.views import SpectacularSwaggerView
 from rest_framework.authtoken.views import obtain_auth_token
-from dashboard.views import dashboard, program_form, eligibility_form
+from dashboard.views import dashboard
+from app.views import eligibility_form
 
 urlpatterns = [
     path("", TemplateView.as_view(template_name="pages/home.html"), name="home"),
@@ -32,13 +33,7 @@ urlpatterns = [
         dashboard,
         name='dashboard',
         kwargs={'allow_direct_user': True},
-     ),
-    path(
-        'program_form',
-        program_form,
-        name='program_form',
-        kwargs={'allow_direct_user': True},
-     ),
+    ),
     path(
         'eligibility_form',
         eligibility_form,
