+++ dashboard/views.py
@@ -26,12 +26,11 @@ from django.shortcuts import render, redirect, reverse
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ValidationError
 
-from .forms import UserForm, AddressFormSet, HouseholdForm, SameAddressForm
+from .forms import UserForm
 from get_your.users.models import User
-from app.models import Address, Household
 from app.backend.address import validate_usps
 from monitor.wrappers import LoggerWrapper
-from ref.models import EligibilityProgram as EligibilityProgramRef
+
 
 # Initialize logger
 log = LoggerWrapper(logging.getLogger(__name__))
@@ -45,139 +44,3 @@ def dashboard(request, **kwargs):
                 "title": "Get FoCo Dashboard",
             },
         )
-
-@login_required(redirect_field_name='auth_next')
-def program_form(request, **kwargs):
-    initial_address_data = []
-
-    address = Address.objects.select_related(
-        'eligibility_address',
-        'mailing_address'
-    ).filter(
-        user_id=request.user.id
-    ).first()
-
-    if address is not None:
-        initial_address_data = address.set_initial_form_data()
-
-    household = Household.objects.prefetch_related(
-        'members'
-    ).filter(
-        user_id=request.user.id
-    ).first()
-
-    user_json_data = {
-        "id": request.user.id,
-        "first_name": request.user.first_name,
-        "last_name": request.user.last_name,
-    }
-
-    if request.method == 'POST':
-        # on post, set form to use posted data to fill form in case of error
-        user_form = UserForm(request.POST, prefix='user', instance=request.user)
-        address_form_set = AddressFormSet(request.POST, initial=initial_address_data)
-        same_address_form = SameAddressForm(request.POST)
-        household_form = HouseholdForm(
-            request.POST,
-            prefix='household',
-            initial={
-                'user': request.user.id
-            },
-            instance=household
-        )
-
-        if user_form.is_valid():
-            user_form.save()
-
-        if household_form.is_valid():
-            household_form.save()
-
-        if address_form_set.is_valid():
-            address_info_for_db = []
-            for address_form in address_form_set:
-                # there will only ever be two address forms total
-                # in the set. the first form represents eligibility address
-                # and should always have data, but mailing address may be empty
-                if address_form.cleaned_data:
-                    # create ref_address model
-                    new_address = address_form.save()
-                    address_info_for_db.append(new_address.id)
-
-            print('addresses valid')
-            # create app_address info and then associate
-            # eligibility address and, if needed, mailing address
-            Address.objects.update_or_create(
-                user_id = request.user.id,
-                eligibility_address_id = address_info_for_db[0],
-                mailing_address_id = address_info_for_db[1] if len(address_info_for_db) > 1 else address_info_for_db[0],
-            )
-
-            return render(
-                request,
-                'dashboard/dashboard.html',
-                {
-                    "title": "Get FoCo Dashboard",
-                },
-            )
-
-        # if validation fails, return form with input
-        return render(
-            request,
-            'dashboard/program_form.html',
-            {
-                'title': 'Program Form',
-                'user_form': user_form,
-                'address_form_set': address_form_set,
-                'same_address_form': same_address_form,
-                'household_form': household_form,
-                'userJson': user_json_data
-            },
-        )
-
-    # if not a POST request
-    user_form = UserForm(prefix='user', instance=request.user)
-    address_form_set = AddressFormSet(initial=initial_address_data)
-    same_address_form = SameAddressForm()
-    household_form = HouseholdForm(
-        prefix='household',
-        initial={
-            'user': request.user.id
-        },
-        instance=household
-    )
-
-    return render(
-            request,
-            'dashboard/program_form.html',
-            {
-                'title': 'Program Form',
-                'user_form': user_form,
-                'address_form_set': address_form_set,
-                'same_address_form': same_address_form,
-                'household_form': household_form,
-                'userJson': user_json_data
-            },
-        )
-
-@login_required(redirect_field_name='auth_next')
-def eligibility_form(request, **kwargs):
-    if request.method == "POST":
-        # Handle the program ID and file upload
-        # Return to some other page, with a sucess mesage
-        pass
-    else:
-        ami_30_programs = EligibilityProgramRef.objects.filter(is_active=True).filter(ami_threshold=.3).order_by(
-                'friendly_name').values_list('friendly_name', flat=True)
-
-        ami_60_programs = EligibilityProgramRef.objects.filter(is_active=True).filter(ami_threshold=.6).order_by(
-                'friendly_name').values_list('friendly_name', flat=True)
-
-        return render(
-            request,
-            'dashboard/eligibility_form.html',
-            {
-                'title': 'Program Form',
-                'ami30s': ami_30_programs,
-                'ami60s': ami_60_programs,
-            }
-    )
